[Java][library: vertx] Add default value and required parameter support to vertx server temp… (#7410)
* Add default value and required parameter support to vertx server templates. Fix for #7409.

* Added missing serviceId declaration to workaround #allParams section clearing the vendorExtensions map

+++ samples/server/petstore/java-vertx/rx/src/main/java/io/swagger/server/api/verticle/UserApiVerticle.java
@@ -44,7 +44,14 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for createUser
         vertx.eventBus().<JsonObject> consumer(CREATEUSER_SERVICE_ID).handler(message -> {
             try {
-                User body = Json.mapper.readValue(message.body().getJsonObject("body").encode(), User.class);
+                // Workaround for #allParams section clearing the vendorExtensions map
+                String serviceId = "createUser";
+                JsonObject bodyParam = message.body().getJsonObject("body");
+                if (bodyParam == null) {
+                    manageError(message, new MainApiException(400, "body is required"), serviceId);
+                    return;
+                }
+                User body = Json.mapper.readValue(bodyParam.encode(), User.class);
                 service.createUser(body).subscribe(
                     () -> {
                         message.reply(null);
@@ -61,8 +68,15 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for createUsersWithArrayInput
         vertx.eventBus().<JsonObject> consumer(CREATEUSERSWITHARRAYINPUT_SERVICE_ID).handler(message -> {
             try {
-                List<User> body = Json.mapper.readValue(message.body().getJsonArray("body").encode(),
-                        Json.mapper.getTypeFactory().constructCollectionType(List.class, User.class));
+                // Workaround for #allParams section clearing the vendorExtensions map
+                String serviceId = "createUsersWithArrayInput";
+                JsonArray bodyParam = message.body().getJsonArray("body");
+                if(bodyParam == null) {
+                    manageError(message, new MainApiException(400, "body is required"), serviceId);
+                    return;
+                }
+                List<User> body = Json.mapper.readValue(bodyParam.encode(),
+                    Json.mapper.getTypeFactory().constructCollectionType(List.class, User.class));
                 service.createUsersWithArrayInput(body).subscribe(
                     () -> {
                         message.reply(null);
@@ -79,8 +93,15 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for createUsersWithListInput
         vertx.eventBus().<JsonObject> consumer(CREATEUSERSWITHLISTINPUT_SERVICE_ID).handler(message -> {
             try {
-                List<User> body = Json.mapper.readValue(message.body().getJsonArray("body").encode(),
-                        Json.mapper.getTypeFactory().constructCollectionType(List.class, User.class));
+                // Workaround for #allParams section clearing the vendorExtensions map
+                String serviceId = "createUsersWithListInput";
+                JsonArray bodyParam = message.body().getJsonArray("body");
+                if(bodyParam == null) {
+                    manageError(message, new MainApiException(400, "body is required"), serviceId);
+                    return;
+                }
+                List<User> body = Json.mapper.readValue(bodyParam.encode(),
+                    Json.mapper.getTypeFactory().constructCollectionType(List.class, User.class));
                 service.createUsersWithListInput(body).subscribe(
                     () -> {
                         message.reply(null);
@@ -97,7 +118,14 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for deleteUser
         vertx.eventBus().<JsonObject> consumer(DELETEUSER_SERVICE_ID).handler(message -> {
             try {
-                String username = message.body().getString("username");
+                // Workaround for #allParams section clearing the vendorExtensions map
+                String serviceId = "deleteUser";
+                String usernameParam = message.body().getString("username");
+                if(usernameParam == null) {
+                    manageError(message, new MainApiException(400, "username is required"), serviceId);
+                    return;
+                }
+                String username = usernameParam;
                 service.deleteUser(username).subscribe(
                     () -> {
                         message.reply(null);
@@ -114,7 +142,14 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for getUserByName
         vertx.eventBus().<JsonObject> consumer(GETUSERBYNAME_SERVICE_ID).handler(message -> {
             try {
-                String username = message.body().getString("username");
+                // Workaround for #allParams section clearing the vendorExtensions map
+                String serviceId = "getUserByName";
+                String usernameParam = message.body().getString("username");
+                if(usernameParam == null) {
+                    manageError(message, new MainApiException(400, "username is required"), serviceId);
+                    return;
+                }
+                String username = usernameParam;
                 service.getUserByName(username).subscribe(
                     result -> {
                         message.reply(new JsonObject(Json.encode(result)).encodePrettily());
@@ -131,8 +166,20 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for loginUser
         vertx.eventBus().<JsonObject> consumer(LOGINUSER_SERVICE_ID).handler(message -> {
             try {
-                String username = message.body().getString("username");
-                String password = message.body().getString("password");
+                // Workaround for #allParams section clearing the vendorExtensions map
+                String serviceId = "loginUser";
+                String usernameParam = message.body().getString("username");
+                if(usernameParam == null) {
+                    manageError(message, new MainApiException(400, "username is required"), serviceId);
+                    return;
+                }
+                String username = usernameParam;
+                String passwordParam = message.body().getString("password");
+                if(passwordParam == null) {
+                    manageError(message, new MainApiException(400, "password is required"), serviceId);
+                    return;
+                }
+                String password = passwordParam;
                 service.loginUser(username, password).subscribe(
                     result -> {
                         message.reply(new JsonObject(Json.encode(result)).encodePrettily());
@@ -149,6 +196,8 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for logoutUser
         vertx.eventBus().<JsonObject> consumer(LOGOUTUSER_SERVICE_ID).handler(message -> {
             try {
+                // Workaround for #allParams section clearing the vendorExtensions map
+                String serviceId = "logoutUser";
                 service.logoutUser().subscribe(
                     () -> {
                         message.reply(null);
@@ -165,8 +214,20 @@ public class UserApiVerticle extends AbstractVerticle {
         //Consumer for updateUser
         vertx.eventBus().<JsonObject> consumer(UPDATEUSER_SERVICE_ID).handler(message -> {
             try {
-                String username = message.body().getString("username");
-                User body = Json.mapper.readValue(message.body().getJsonObject("body").encode(), User.class);
+                // Workaround for #allParams section clearing the vendorExtensions map
+                String serviceId = "updateUser";
+                String usernameParam = message.body().getString("username");
+                if(usernameParam == null) {
+                    manageError(message, new MainApiException(400, "username is required"), serviceId);
+                    return;
+                }
+                String username = usernameParam;
+                JsonObject bodyParam = message.body().getJsonObject("body");
+                if (bodyParam == null) {
+                    manageError(message, new MainApiException(400, "body is required"), serviceId);
+                    return;
+                }
+                User body = Json.mapper.readValue(bodyParam.encode(), User.class);
                 service.updateUser(username, body).subscribe(
                     () -> {
                         message.reply(null);
